+++ Manager/SceneManager.cpp
@@ -1,0 +1,10 @@
+#include "SceneManager.h"
+
+SceneManager::SceneManager(SceneType s_type)
+	:s_type(s_type)
+{
+}
+
+SceneManager::~SceneManager()
+{
+}
