+++ Manager/SceneManager.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "../Scene/Scene.h"
+
+class SceneManager {
+private:
+	SceneType s_type;
+
+
+
+public:
+	SceneManager(SceneType s_type);
+
+	~SceneManager();
+};
