+++ sfml-template.vcxproj
@@ -148,7 +148,24 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Framework\GameFramework.cpp" />
+    <ClCompile Include="GameObject\ImgObj.cpp" />
+    <ClCompile Include="GameObject\Player.cpp" />
+    <ClCompile Include="Manager\SceneManager.cpp" />
+    <ClCompile Include="Scene\Scene.cpp" />
+    <ClCompile Include="Scene\StageScene.cpp" />
     <ClCompile Include="sfml-template.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Framework\GameFramework.h" />
+    <ClInclude Include="GameObject\ImgObj.h" />
+    <ClInclude Include="GameObject\Player.h" />
+    <ClInclude Include="Manager\Resource.h" />
+    <ClInclude Include="Manager\ResourceManager.h" />
+    <ClInclude Include="Manager\SceneManager.h" />
+    <ClInclude Include="Manager\Singleton.h" />
+    <ClInclude Include="Scene\Scene.h" />
+    <ClInclude Include="Scene\StageScene.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
